pyproject.toml
[tool.poetry]
name = "chopin"
version = "0.1.0-0"
description = "chopin: Compose and play with Spotify playlists"
authors = ["pstfni"]
packages = [{include="chopin"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
numpy = "^1.6.0"
tqdm = "^4.64.0"
emoji = "~1.6.3"
"ruamel.yaml" = "^0.17"
spotipy = "^2.19"
pydantic = "^1.9"
python-dotenv = "^0.19.2"
coloredlogs = "^15.0"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]
pytest-sugar = "*"
ruff = "*"
pandas = "*"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.3"
mkdocstrings = "^0.20"
mkdocstrings-python = "^0.8"
mkdocs-material = "^8.5"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
scikit-learn = "^1.2.1"

[tool.ruff]
line-length = 120
exclude = ['.venv']
ignore = ["D101", "D102", "D106", "D107", "N805"]
select = ["F", "E", "W", "I", "N", "D", "UP", "TRY200", "PERF", "RUF"]

[tool.ruff.per-file-ignores]
"**/__init__.py" = ['F401']
"**/tests/**" = ["S", "D", "N"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.poetry.scripts]
composer = "chopin.entrypoints.composer:main"
queuer = "chopin.entrypoints.queuer:main"
describer = "chopin.entrypoints.describer:main"
shuffler = "chopin.entrypoints.shuffler:main"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
